Extract project display logic to composable
Add useProjectDisplay composable to centralize computed properties for projects (progressPercent, completedTaskCount, taskSummary, createdDate) and reduce duplication. Update ProjectCard.vue and ProjectListRow.vue to use the new composable. Also fix status label typo from "Inprogress" to "In Progress" in StatusBadge.vue and the status options in pages/workspace/[id]/projects.vue to keep UI text consistent.
